refactor: extract core extension api policy hooks
