+++ addons/smart_construction_core/core_extension.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 import logging
-from typing import Any, Dict, List
 
-from odoo import fields
 from odoo.addons.smart_core.core.delivery_menu_defaults import register_current_project_scope_model
 from odoo.addons.smart_core.core.project_context import (
     register_business_scope_exempt_model,
@@ -121,21 +119,33 @@ from odoo.addons.smart_construction_core.core_extension_services import (
 from odoo.addons.smart_construction_core.core_extension_projected_contracts import (
     smart_core_finalize_projected_contract_data,
 )
+from odoo.addons.smart_construction_core.core_extension_api_policy import (
+    get_api_data_create_execution_policy_contribution,
+    get_api_data_mutation_policy_contribution,
+    get_api_data_unlink_allowed_model_contributions,
+    get_api_data_write_allowlist_contributions,
+    get_file_download_allowed_model_contributions,
+    get_file_upload_allowed_model_contributions,
+    get_intent_permission_model_acl_policy_contribution,
+    get_model_code_mapping_contributions,
+    get_server_action_window_map_contributions,
+    smart_core_api_data_create_execution_policy,
+    smart_core_api_data_mutation_policy,
+    smart_core_api_data_search_fields,
+    smart_core_api_data_unlink_allowed_models,
+    smart_core_api_data_write_allowlist,
+    smart_core_file_download_allowed_models,
+    smart_core_file_download_auth_subject,
+    smart_core_file_upload_allowed_models,
+    smart_core_intent_permission_model_acl_policy,
+    smart_core_legacy_visible_business_column_labels,
+    smart_core_model_code_mapping,
+    smart_core_server_action_window_map,
+)
 from odoo.addons.smart_construction_core.core_extension_policy_catalog import (
-    API_DATA_MUTATION_POLICIES,
-    API_DATA_WRITE_ALLOWLIST,
-    DRAFT_DELETE_ALLOWED_STATES,
-    FILE_ATTACHMENT_ALLOWED_LEGACY_MODEL_PREFIXES,
-    FILE_ATTACHMENT_ALLOWED_MODEL_EXACT,
-    FILE_ATTACHMENT_ALLOWED_MODEL_PREFIXES,
-    FILE_ATTACHMENT_EXCLUDED_MODEL_PREFIXES,
-    FILE_DOWNLOAD_ALLOWED_MODELS,
-    FILE_UPLOAD_ALLOWED_MODELS,
     CRITICAL_SCENE_TARGET_OVERRIDES,
     CRITICAL_SCENE_TARGET_ROUTE_OVERRIDES,
     INDUSTRY_CREATE_FIELD_FALLBACKS,
-    LEGACY_VISIBLE_BUSINESS_COLUMN_LABELS_BY_MODEL,
-    MODEL_CODE_MAPPING,
     NAV_ACTION_SCENE_MAP,
     NAV_MENU_SCENE_MAP,
     NAV_MODEL_VIEW_SCENE_MAP,
@@ -143,7 +153,6 @@ from odoo.addons.smart_construction_core.core_extension_policy_catalog import (
     ROLE_GROUPS_EXPLICIT,
     ROLE_PRECEDENCE,
     ROLE_SURFACE_OVERRIDES,
-    SERVER_ACTION_WINDOW_MAP,
 )
 
 _logger = logging.getLogger(__name__)
@@ -697,232 +706,6 @@ register_legacy_standard_list_profile({
     "strict_columns": True,
 })
 
-def _state_unlink_policy(
-    model_name: str,
-    business_label: str,
-    allowed_states=DRAFT_DELETE_ALLOWED_STATES,
-    state_field: str = "state",
-):
-    return {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "policy_kind": "state_limited_business_document",
-        "state_field": state_field,
-        "allowed_states": list(allowed_states),
-        "reason_code": "DRAFT_BUSINESS_DOCUMENT_DELETE_ALLOWED",
-        "message": f"允许删除未形成业务事实的{business_label}；仅限草稿/取消等未提交状态，并继续受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    }
-
-
-API_DATA_DRAFT_UNLINK_POLICIES = {
-    "construction.contract": _state_unlink_policy("construction.contract", "合同记录"),
-    "construction.contract.income": _state_unlink_policy("construction.contract.income", "收入合同"),
-    "construction.contract.expense": _state_unlink_policy("construction.contract.expense", "支出合同"),
-    "payment.request": _state_unlink_policy("payment.request", "付款申请"),
-    "sc.general.contract": _state_unlink_policy("sc.general.contract", "综合合同"),
-    "sc.expense.claim": _state_unlink_policy("sc.expense.claim", "费用与保证金单据"),
-    "sc.financing.loan": _state_unlink_policy("sc.financing.loan", "融资借款单据"),
-    "sc.invoice.registration": _state_unlink_policy("sc.invoice.registration", "发票登记单据"),
-    "sc.payment.execution": _state_unlink_policy("sc.payment.execution", "付款执行单"),
-    "sc.receipt.income": _state_unlink_policy("sc.receipt.income", "收款收入登记"),
-    "sc.fund.account.operation": _state_unlink_policy("sc.fund.account.operation", "资金账户操作单"),
-    "sc.self.funding.registration": _state_unlink_policy("sc.self.funding.registration", "自筹资金登记"),
-    "sc.tax.deduction.registration": _state_unlink_policy("sc.tax.deduction.registration", "税票抵扣登记"),
-    "sc.settlement.order": _state_unlink_policy("sc.settlement.order", "结算单"),
-    "sc.settlement.adjustment": _state_unlink_policy("sc.settlement.adjustment", "结算调整单"),
-    "project.material.plan": _state_unlink_policy("project.material.plan", "材料计划"),
-    "sc.material.purchase.request": _state_unlink_policy("sc.material.purchase.request", "材料采购申请"),
-    "sc.material.acceptance": _state_unlink_policy("sc.material.acceptance", "材料验收单"),
-    "sc.material.inbound": _state_unlink_policy("sc.material.inbound", "材料入库单"),
-    "sc.material.outbound": _state_unlink_policy("sc.material.outbound", "材料出库单"),
-    "sc.material.rfq": _state_unlink_policy("sc.material.rfq", "材料询比价"),
-    "sc.material.settlement": _state_unlink_policy("sc.material.settlement", "材料结算单"),
-    "sc.material.rental.plan": _state_unlink_policy("sc.material.rental.plan", "材料租赁计划"),
-    "sc.material.rental.order": _state_unlink_policy("sc.material.rental.order", "材料租赁订单"),
-    "sc.material.rental.settlement": _state_unlink_policy("sc.material.rental.settlement", "材料租赁结算"),
-    "sc.labor.plan": _state_unlink_policy("sc.labor.plan", "劳务计划"),
-    "sc.labor.request": _state_unlink_policy("sc.labor.request", "劳务申请"),
-    "sc.labor.usage": _state_unlink_policy("sc.labor.usage", "劳务使用记录"),
-    "sc.labor.settlement": _state_unlink_policy("sc.labor.settlement", "劳务结算"),
-    "sc.labor.price": _state_unlink_policy("sc.labor.price", "劳务价格单"),
-    "sc.equipment.plan": _state_unlink_policy("sc.equipment.plan", "设备计划"),
-    "sc.equipment.request": _state_unlink_policy("sc.equipment.request", "设备申请"),
-    "sc.equipment.usage": _state_unlink_policy("sc.equipment.usage", "设备使用记录"),
-    "sc.equipment.settlement": _state_unlink_policy("sc.equipment.settlement", "设备结算"),
-    "sc.equipment.price": _state_unlink_policy("sc.equipment.price", "设备价格单"),
-    "sc.safety.plan": _state_unlink_policy("sc.safety.plan", "安全方案"),
-    "sc.safety.disclosure": _state_unlink_policy("sc.safety.disclosure", "安全交底"),
-    "sc.safety.issue": _state_unlink_policy("sc.safety.issue", "安全问题"),
-    "sc.safety.patrol.task": _state_unlink_policy("sc.safety.patrol.task", "安全巡检任务"),
-    "sc.quality.issue": _state_unlink_policy("sc.quality.issue", "质量问题"),
-    "sc.quality.rectification": _state_unlink_policy(
-        "sc.quality.rectification",
-        "质量整改记录",
-        ("submitted", "rectifying", "rechecking", "cancel"),
-        state_field="issue_state",
-    ),
-    "sc.quality.recheck": _state_unlink_policy(
-        "sc.quality.recheck",
-        "质量复验记录",
-        ("rectifying", "rechecking", "cancel"),
-        state_field="issue_state",
-    ),
-    "sc.safety.rectification": _state_unlink_policy(
-        "sc.safety.rectification",
-        "安全整改记录",
-        ("submitted", "rectifying", "rechecking", "cancel"),
-        state_field="issue_state",
-    ),
-    "sc.safety.recheck": _state_unlink_policy(
-        "sc.safety.recheck",
-        "安全复验记录",
-        ("rectifying", "rechecking", "cancel"),
-        state_field="issue_state",
-    ),
-    "sc.construction.diary": _state_unlink_policy("sc.construction.diary", "施工日志"),
-    "project.progress.entry": _state_unlink_policy("project.progress.entry", "进度填报"),
-    "project.risk.action": _state_unlink_policy("project.risk.action", "风险措施"),
-    "sc.plan": _state_unlink_policy("sc.plan", "项目计划"),
-    "sc.plan.line": _state_unlink_policy("sc.plan.line", "项目计划明细"),
-    "sc.plan.version": _state_unlink_policy("sc.plan.version", "计划版本"),
-    "sc.plan.report": _state_unlink_policy("sc.plan.report", "计划上报"),
-    "tender.bid": _state_unlink_policy("tender.bid", "投标主单", ("prepare", "estimating")),
-    "tender.doc.purchase": _state_unlink_policy("tender.doc.purchase", "投标文件购买申请"),
-    "tender.doc.review": _state_unlink_policy("tender.doc.review", "投标文件审查"),
-    "tender.guarantee": _state_unlink_policy("tender.guarantee", "投标保证金"),
-}
-API_DATA_UNLINK_POLICIES = {
-    "construction.contract": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理合同记录；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "construction.contract.income": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理收入合同记录；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "construction.contract.expense": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理支出合同记录；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "hr.department": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理组织部门；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "payment.request": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理付款申请；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "payment.request.line": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理付款申请明细；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "project.cost.code": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理成本科目；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "project.dictionary": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理业务字典；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "project.task": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许删除任务记录；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "project.tags": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "RELATION_MAINTENANCE_DELETE_ALLOWED",
-        "message": "允许删除项目标签等关系维护数据；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "res.partner": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理客户/供应商资料；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "sc.approval.policy": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理审批策略；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "sc.approval.step": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理审批步骤；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "sc.document.admin.document": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理行政档案；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "sc.hr.payroll.document": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理薪酬档案；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "sc.office.admin.document": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理办公行政资料；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "sc.project.stage.requirement.item": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理阶段要求；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-    "sc.supplier.type": {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "DELETE_POLICY_ALLOWED",
-        "message": "允许业务配置管理员整理供应商类型；仍受模型 ACL 与记录规则约束。",
-        "source": "smart_construction_core",
-    },
-}
-API_DATA_UNLINK_POLICIES.update(API_DATA_DRAFT_UNLINK_POLICIES)
-API_DATA_UNLINK_ALLOWED_MODELS = list(API_DATA_UNLINK_POLICIES)
-
 def smart_core_identity_profile(env):
     return {
         "role_surface_map": ROLE_SURFACE_OVERRIDES,
@@ -970,155 +753,6 @@ def smart_core_critical_scene_target_overrides(env):
 
 def smart_core_critical_scene_target_route_overrides(env):
     return dict(CRITICAL_SCENE_TARGET_ROUTE_OVERRIDES)
-
-
-def get_server_action_window_map_contributions(env):
-    return dict(SERVER_ACTION_WINDOW_MAP)
-
-
-def get_file_upload_allowed_model_contributions(env):
-    return sorted(set(FILE_UPLOAD_ALLOWED_MODELS) | _business_attachment_allowed_models(env))
-
-
-def get_file_download_allowed_model_contributions(env):
-    return sorted(set(FILE_DOWNLOAD_ALLOWED_MODELS) | _business_attachment_allowed_models(env))
-
-
-def _business_attachment_allowed_models(env):
-    out = set()
-    try:
-        models = env["ir.model"].sudo().search([])
-    except Exception:
-        return out
-    for row in models:
-        model_name = str(row.model or "").strip()
-        if not model_name:
-            continue
-        legacy_attachment_model = model_name.startswith(FILE_ATTACHMENT_ALLOWED_LEGACY_MODEL_PREFIXES)
-        if (
-            model_name not in FILE_ATTACHMENT_ALLOWED_MODEL_EXACT
-            and not legacy_attachment_model
-            and model_name.startswith(FILE_ATTACHMENT_EXCLUDED_MODEL_PREFIXES)
-        ):
-            continue
-        if (
-            model_name not in FILE_ATTACHMENT_ALLOWED_MODEL_EXACT
-            and not legacy_attachment_model
-            and not model_name.startswith(FILE_ATTACHMENT_ALLOWED_MODEL_PREFIXES)
-        ):
-            continue
-        if model_name not in env:
-            continue
-        try:
-            if getattr(env[model_name], "_transient", False) or getattr(env[model_name], "_abstract", False):
-                continue
-        except Exception:
-            continue
-        out.add(model_name)
-    return out
-
-
-def get_api_data_write_allowlist_contributions(env):
-    return {
-        str(model_name): list(field_names)
-        for model_name, field_names in API_DATA_WRITE_ALLOWLIST.items()
-    }
-
-
-def get_api_data_mutation_policy_contribution(env, model_name: str, op: str):
-    policy = API_DATA_MUTATION_POLICIES.get(str(model_name or "").strip())
-    if not isinstance(policy, dict):
-        return {"allowed": True, "reason_code": "OK", "source": "smart_construction_core"}
-    allowed_ops = {
-        str(item or "").strip().lower()
-        for item in (policy.get("allowed_ops") or [])
-        if str(item or "").strip()
-    }
-    normalized_op = str(op or "").strip().lower()
-    if allowed_ops and normalized_op not in allowed_ops:
-        return {"allowed": True, "reason_code": "OK", "source": "smart_construction_core"}
-    out = dict(policy)
-    out["op"] = normalized_op
-    out["model"] = str(model_name or "").strip()
-    return out
-
-
-def _is_contract_tax_rate_quick_create(env, vals: dict) -> bool:
-    safe_vals = vals if isinstance(vals, dict) else {}
-    if (
-        safe_vals.get("type_tax_use") == "none"
-        and safe_vals.get("amount_type") == "percent"
-        and safe_vals.get("price_include") is False
-        and safe_vals.get("tax_group_id")
-    ):
-        try:
-            group = env["account.tax.group"].sudo().browse(int(safe_vals.get("tax_group_id") or 0)).exists()
-        except Exception:
-            group = env["account.tax.group"].browse()
-        if group and group.name == "合同税率":
-            return True
-    return False
-
-
-def get_intent_permission_model_acl_policy_contribution(env, intent_name: str, model_name: str, access_mode: str, params: dict):
-    if (
-        str(intent_name or "").strip() == "api.data"
-        and str(model_name or "").strip() == "account.tax"
-        and str(access_mode or "").strip() == "create"
-    ):
-        raw_params = params if isinstance(params, dict) else {}
-        payload = raw_params.get("params") if isinstance(raw_params.get("params"), dict) else raw_params
-        if isinstance(raw_params.get("payload"), dict):
-            payload = raw_params.get("payload")
-        vals = payload.get("vals") or payload.get("values") if isinstance(payload, dict) else {}
-        if _is_contract_tax_rate_quick_create(env, vals if isinstance(vals, dict) else {}):
-            return {
-                "skip_model_acl": True,
-                "reason_code": "CONTRACT_TAX_RATE_QUICK_CREATE",
-                "source": "smart_construction_core",
-            }
-    return {"skip_model_acl": False, "source": "smart_construction_core"}
-
-
-def get_api_data_create_execution_policy_contribution(env, model_name: str, vals: dict, ctx: dict, params: dict):
-    model = str(model_name or "").strip()
-    safe_vals = vals if isinstance(vals, dict) else {}
-    if model != "account.tax":
-        return {"sudo": False, "source": "smart_construction_core"}
-    if _is_contract_tax_rate_quick_create(env, safe_vals):
-        return {
-            "allowed": True,
-            "sudo": True,
-            "reason_code": "CONTRACT_TAX_RATE_QUICK_CREATE",
-            "source": "smart_construction_core",
-        }
-    return {
-        "allowed": False,
-        "sudo": False,
-        "reason_code": "ACCOUNT_TAX_NATIVE_CREATE_FORBIDDEN",
-        "message": "税率只能通过合同税率百分比快建，不能维护原生会计税种。",
-        "source": "smart_construction_core",
-    }
-
-
-def get_api_data_unlink_allowed_model_contributions(env):
-    policies = {
-        str(model_name): dict(policy)
-        for model_name, policy in API_DATA_UNLINK_POLICIES.items()
-    }
-    policies["project.project"] = {
-        "allowed": True,
-        "delete_mode": "unlink",
-        "reason_code": "PROJECT_MASTER_DELETE_ALLOWED",
-        "message": "允许删除无业务依赖的项目主数据；继续受模型 ACL、记录规则与项目依赖阻断约束。",
-        "source": "smart_construction_core",
-        "dependency_guard": "project.project._raise_project_unlink_blockers",
-    }
-    return policies
-
-
-def get_model_code_mapping_contributions(env):
-    return dict(MODEL_CODE_MAPPING)
 
 
 def smart_core_register(registry):
@@ -1321,102 +955,3 @@ def smart_core_extend_system_init(data, env, user):
         ext_facts[module_key] = dict(facts_payload)
     data["ext_facts"] = ext_facts
     return apply_system_init_profile_overrides(data, ext_facts)
-
-
-def smart_core_server_action_window_map(env):
-    return get_server_action_window_map_contributions(env)
-
-
-def smart_core_file_upload_allowed_models(env):
-    return get_file_upload_allowed_model_contributions(env)
-
-
-def smart_core_file_download_allowed_models(env):
-    return get_file_download_allowed_model_contributions(env)
-
-
-def smart_core_file_download_auth_subject(env, attachment, current_subject):
-    del current_subject
-    try:
-        if "payment.request" not in env:
-            return None
-        parent_request = env["payment.request"].sudo().search(
-            [("attachment_ids", "in", attachment.id)],
-            limit=1,
-        )
-    except Exception:
-        return None
-    if not parent_request:
-        return None
-    return {"model": "payment.request", "res_id": parent_request.id}
-
-
-def smart_core_legacy_visible_business_column_labels(env):
-    del env
-    return LEGACY_VISIBLE_BUSINESS_COLUMN_LABELS_BY_MODEL
-
-
-def smart_core_api_data_write_allowlist(env):
-    return get_api_data_write_allowlist_contributions(env)
-
-
-def smart_core_api_data_mutation_policy(env, model_name: str, op: str):
-    return get_api_data_mutation_policy_contribution(env, model_name, op)
-
-
-def smart_core_intent_permission_model_acl_policy(env, intent_name: str, model_name: str, access_mode: str, params: dict):
-    return get_intent_permission_model_acl_policy_contribution(env, intent_name, model_name, access_mode, params)
-
-
-def smart_core_api_data_create_execution_policy(env, model_name: str, vals: dict, ctx: dict, params: dict):
-    return get_api_data_create_execution_policy_contribution(env, model_name, vals, ctx, params)
-
-
-def smart_core_api_data_unlink_allowed_models(env):
-    return get_api_data_unlink_allowed_model_contributions(env)
-
-
-def smart_core_api_data_search_fields(env, model_name: str):
-    try:
-        from .models.support.p1_daily_business_visible_alias_fields import (
-            LABEL_SOURCE_OVERRIDES,
-            MODEL_LABEL_SOURCE_OVERRIDES,
-            P1_ALIAS_COMPAT_LABELS,
-            P1_ALIAS_LABELS,
-        )
-        from .models.support.user_confirmed_formal_visible_fields import USER_CONFIRMED_FORMAL_VISIBLE_FIELDS
-    except Exception:
-        return []
-
-    model_name = str(model_name or "").strip()
-    labels = []
-    for label in list(P1_ALIAS_LABELS.get(model_name) or []) + list(P1_ALIAS_COMPAT_LABELS.get(model_name) or []):
-        value = str(label or "").strip()
-        if value and value not in labels:
-            labels.append(value)
-    for entry in USER_CONFIRMED_FORMAL_VISIBLE_FIELDS.get(model_name) or []:
-        label = str((entry or {}).get("label") or "").strip()
-        if label and label not in labels:
-            labels.append(label)
-    model_overrides = MODEL_LABEL_SOURCE_OVERRIDES.get(model_name) or {}
-    for label in model_overrides:
-        value = str(label or "").strip()
-        if value and value not in labels:
-            labels.append(value)
-    names = []
-    for label in labels:
-        for field_name in list(model_overrides.get(label) or []) + list(LABEL_SOURCE_OVERRIDES.get(label) or []):
-            value = str(field_name or "").strip()
-            if value and value not in names:
-                names.append(value)
-    if env is None:
-        return names
-    try:
-        model_fields = getattr(env[model_name], "_fields", {}) or {}
-    except Exception:
-        return names
-    return [field_name for field_name in names if field_name in model_fields]
-
-
-def smart_core_model_code_mapping(env):
-    return get_model_code_mapping_contributions(env)
